Implemented asynchronous mutator API
-C++ client done (Future and TableMutatorAsync, along with tests)
-ThriftAPI changes implemented
-Thrift tests updated

+++ src/cc/Hypertable/Lib/TableMutatorAsyncHandler.cc
@@ -37,7 +37,7 @@ using namespace Serialization;
  *
  */
 TableMutatorAsyncHandler::TableMutatorAsyncHandler(TableMutatorAsync *mutator,
-    size_t scatter_buffer) : ApplicationHandler(0),
+    uint32_t scatter_buffer) : ApplicationHandler(0),
     m_mutator(mutator), m_scatter_buffer(scatter_buffer) {
 }
 
@@ -47,9 +47,12 @@ TableMutatorAsyncHandler::TableMutatorAsyncHandler(TableMutatorAsync *mutator,
 void TableMutatorAsyncHandler::run() {
   // The scatter buffer will get destroyed when the TableMutator releases it
   // and this method goes out of scope
+  //HT_INFO_OUT << "[bibble] Running TableMutatorAsyncHandler " << std::hex << this << HT_END;
   TableMutatorAsyncScatterBufferPtr buffer =
       m_mutator->get_outstanding_buffer(m_scatter_buffer);
   HT_ASSERT(buffer);
   buffer->finish();
+  //HT_INFO_OUT << "[bibble] TableMutatorAsyncHandler " << std::hex << this << " completed"
+  //    << HT_END;
 }
